+++ CHANGELOG.md
@@ -97,6 +97,7 @@ v2.0.0
 - Changed object iteration callbacks to be value first.
 - Renamed `String#each` to `String#forEach`.
 - Changed `min/max/least/most` args to put the callback at the end.
+- Removed `String#paragraphs`.
 
 
 v1.4.2